Fix for a crash, and addition of some backend code.
Crash involved attempting to hide a mdi window after the child window
had been disposed of. Apparently Qt automatically takes care of
disposing the mdi window, so I don't need to do it.

Backend code is to allow for multiple display types of data, based on
which type is clicked to be displayed in the ems info window