Bunch of refactoring and other code cleanup and warning removal
Added a sendPacket function to freeemscomms.h/cpp, which should
facilitate a OAOO approach. It greatly reduced the number of repeated
code within the thread loop. It also allowed for significant
expandability in the future in terms of types of arguments to send
along with messages and how they are parsed into byte arrays.