Added read only and config meta data, change in how meta-data calc is loaded
Meta-data calc variables are now lists of pairs of "type" and "value",
which allows for them to be in order, rather than in a map where order is
not garunteed.
Readonly and config meta data is there, but the provisions for parsing
the data from memory is not yet implemented.