Fix for table export, removed dirty hack since QJson isn't broken
I thought that QJson was broken, however it turns out if you append
one QVariantList to another QVariantList, it dosen't add it as a standard
QVariant, it adds each item as their own QVariant. This is fixed by
casting the QVariantList to a QVariant, THEN adding it to the list.