TableView3D: Rapid/holddown keypresses on 3d table now act appropriatly
A key-holddown will now send 3 updates per second, and one final updates
once the key is released. This gives the UI a fluid feel of scrolling, while
not slamming on the COM link. This does mean that the UI is technically out
of sync with the ECU for 1/3 of a second, but that is acceptable to retain a
good UI experience